% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimated_node_dates.R
\name{addEstimatedNodeDates}
\alias{addEstimatedNodeDates}
\title{Add estiamted node dates to tree_tibble}
\usage{
addEstimatedNodeDates(tree_and_sequences)
}
\arguments{
\item{tree_and_sequences}{a list containing `tree` and `tree_tibble`

Two types of node date are added: the collection date of the nearest tip, and the collection date of the nearest _descendant_ tip.

To do:
- Node dating with TimeTree and/or Chronumental
- Earliest date of any descendant tip (but this is very sensitive to misplaced sequences)}
}
\description{
Add estiamted node dates to tree_tibble
}
