% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimated_node_dates.R
\name{addNearestDescendantNodeDates}
\alias{addNearestDescendantNodeDates}
\title{Add estimated node dates to tree_tibble}
\usage{
addNearestDescendantNodeDates(tree_and_sequences)
}
\arguments{
\item{tree_and_sequences}{a list containing \code{tree} and \code{tree_tibble}

Two types of node date are added: the collection date of the nearest tip, and the collection date of the nearest \emph{descendant} tip.}
}
\description{
Add estimated node dates to tree_tibble
}
