% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence-package.R
\docType{package}
\name{convergence-package}
\alias{convergence}
\alias{convergence-package}
\title{convergence: Convergence functions}
\description{
Functions for calculating convergence ratios.
}
\author{
\strong{Maintainer}: Sam Turner \email{sat65@canm.ac.uk}

}
\keyword{internal}
