% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{findBranch}
\alias{findBranch}
\title{Find branch with mutations}
\usage{
findBranch(tree_tibble, target_aa_substitutions, target_nt_mutations)
}
\arguments{
\item{tree_tibble}{tree_tibble}

\item{target_aa_substitutions}{aa substitutions which the branch should have}

\item{target_nt_mutations}{nt mutations which the branch should have}
}
\description{
Finds the branch where all of the target_aa_substitutions and target_nt_mutations are present, and at leaast occurs on the branch. Errors if the number of such branches is not exactly one.
}
