% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clade_sizes.R
\name{getCladeSizeCDFs}
\alias{getCladeSizeCDFs}
\title{Distribution of clade sizes compared to synonymous expectation}
\usage{
getCladeSizeCDFs(
  tree_and_sequences,
  tree_tibble_for_node_search = tree_and_sequences$tree_tibble,
  aa_substitutions,
  syn_nuc_mutations,
  n_bootstraps = 100,
  max_log2_clade_size = 10,
  log2_clade_size_resolution = 0.1
)
}
\arguments{
\item{tree_and_sequences}{...}

\item{tree_tibble_for_node_search}{Perhaps you want to restrict node search to a subset of the tree in tree_and_sequences. Because tree_and_sequences$tree is used to count descendant tips, the "intact" tree_and_sequences (i.e. without tree_and_sequences being subsetted, meaning \code{tree} corresponds exactly to \code{tree_tibble}) must also be passed}

\item{aa_substitutions}{...}

\item{syn_nuc_mutations}{...}

\item{n_bootstraps}{...}

\item{max_log2_clade_size}{...}

\item{log2_clade_size_resolution}{...}
}
\value{
...
}
\description{
Distribution of clade sizes compared to synonymous expectation
}
