% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{getConsensus}
\alias{getConsensus}
\title{Get the consensus sequence from a vector of sequences}
\usage{
getConsensus(sequences)
}
\arguments{
\item{sequences}{a vector of sequences}
}
\description{
Get the consensus sequence from a vector of sequences
}
