% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{getInfoAboutNodes}
\alias{getInfoAboutNodes}
\title{Helper function to get information about nodes in the tree}
\usage{
getInfoAboutNodes(tree_and_sequences, nodes, comparison_sequence)
}
\arguments{
\item{tree_and_sequences}{...}

\item{nodes}{nodes to get info about}

\item{comparison_sequence}{for genetic context information}
}
\description{
Useful for learning more about occurrences of substitutions of interest - e.g. genetic context
}
