% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_aa_ratios.R
\name{getSubstitutionRatios}
\alias{getSubstitutionRatios}
\title{Calculate log convergence ratios}
\usage{
getSubstitutionRatios(
  t_and_s_filtered,
  nuc_rates,
  tree_size_fn,
  sampling_function,
  positions
)
}
\arguments{
\item{t_and_s_filtered}{a list containing \code{tree} and \code{tree_tibble}}

\item{nuc_rates}{from \code{getTreeSizeAndNucRates}}

\item{tree_size_fn}{from \code{getTreeSizeAndNucRates}}

\item{sampling_function}{from \code{getTreeSizeAndNucRates}}

\item{positions}{aa positions for which to calculate log convergence}
}
\description{
Calculate log convergence ratios
}
