% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_size.R
\name{getTreeSizeAndNucRates}
\alias{getTreeSizeAndNucRates}
\title{Get information about a tree required for calculating convergence scores}
\usage{
getTreeSizeAndNucRates(
  tree_and_sequences,
  noise_model,
  n_iter = 2000,
  n_chains = 3,
  excluded_positions = c(),
  use_fitted_rates = T,
  seed = NULL
)
}
\arguments{
\item{tree_and_sequences}{a list containing `tree` and `tree_tibble`

tree_size_ratios are the
tree_size = tree_size,
nuc_rates = nuc_rates,
fast_tree_size_fn = tree_size_fn}
}
\description{
Get information about a tree required for calculating convergence scores
}
