% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tree_functions.R
\name{make_iq_tree}
\alias{make_iq_tree}
\title{Wrapper for IQ-TREE}
\usage{
make_iq_tree(
  sequences,
  wd,
  bootstrap_replicates = NULL,
  num_cores = NULL,
  SH_aLRT_replicates = NULL,
  model = NULL,
  redo = FALSE,
  echo = FALSE,
  ...
)
}
\arguments{
\item{sequences}{DNA alignment to use for phylogenetic analysis. Must be matrix
(i.e., aligned sequences) of class DNAbin}

\item{wd}{Path to working directory. The alignment and IQ-TREE intermediate files
and results will be written here.}

\item{bootstrap_replicates}{Optional; number of ultrafast bootstrap replicates to run.}

\item{num_cores}{Optional; number of cores to use. Set to "AUTO" to determine automatically.}

\item{SH_aLRT_replicates}{Optional; number of SH-aLRT tests to run.}

\item{model}{Optional; specify model. If no model is given, ModelTest will be run
to identify the best model for the data.}

\item{redo}{Logical; should the analysis be redone from scratch if output from
previous runs is present?}

\item{echo}{Logical; should STDERR be written to the screen?}

\item{...}{Other arguments not used by this function but used by
drake for tracking.}
}
\value{
Phylogenetic tree (list of class "phylo")
}
\description{
Wrapper for IQ-TREE
}
