% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASR_functions.R
\name{makeTreeAndSequences}
\alias{makeTreeAndSequences}
\title{Make tree_and_sequences object}
\usage{
makeTreeAndSequences(tree, sequences)
}
\arguments{
\item{tree}{phylo object}

\item{sequences}{sequences dataframe}
}
\description{
Make tree_and_sequences object
}
\details{
The sequences dataframe must contain columns \code{Isolate_unique_identifier} and \code{dna_sequence}. Other columns containing inforamtion about the sequences can also be present. Tip names in the tree in should correspond to the \code{Isolate_unique_identifier} column of sequences.
}
