% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxlike_models.R
\docType{data}
\name{maxlike_models}
\alias{maxlike_models}
\title{Synonymous rate variation models}
\format{
An object of class \code{list} of length 1.
}
\usage{
maxlike_models
}
\description{
Synonymous rate variation models
}
\keyword{datasets}
