% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_models.R
\docType{data}
\name{models}
\alias{models}
\title{Synonymous rate variation models}
\format{
An object of class \code{list} of length 2.
}
\usage{
models
}
\description{
Synonymous rate variation models
}
\keyword{datasets}
