% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reoptimise_tree_and_sequences.R
\name{reoptimiseTreeAndSequences}
\alias{reoptimiseTreeAndSequences}
\title{Reoptimise the tree in a tree_and_sequences object}
\usage{
reoptimiseTreeAndSequences(tree_and_sequences, f)
}
\arguments{
\item{tree_and_sequences}{tree_and_sequences list}

\item{f}{function that takes a named vector of DNA sequences and returns a phylo object}
}
\description{
Reoptimise the tree in a tree_and_sequences object
}
\examples{
\dontrun{
  reoptimiseTreeAndSequences(
    tree_and_sequences,
    function(seqs) {
      t = seqUtils::make_cmaple_tree(
        sequences = c(seqs, outgroup = seqUtils::alaska_232_2015_nts),
        tree_path = "tmp.nwk",
        cmaple_path = "/path/to/bin",
        keep_files = c(),
        out_sequence = "outgroup"
      )
      ape::drop.tip(t, "outgroup")
    }
  )
}

}
