% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{trimTreeAndSequences}
\alias{trimTreeAndSequences}
\title{Trim tree and sequences to descendants of a branch}
\usage{
trimTreeAndSequences(tree_and_sequences, branch)
}
\arguments{
\item{tree_and_sequences}{tree_and_sequences list}

\item{branch}{branch to keep descendants of}
}
\description{
Trim tree and sequences to descendants of a branch
}
