% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASR_functions.R
\name{addASRusher}
\alias{addASRusher}
\title{Add ancestral state reconstruction to \code{tree_and_sequences}}
\usage{
addASRusher(tree_and_sequences, aa_ref, nuc_ref, usher_path = NULL)
}
\arguments{
\item{tree_and_sequences}{list with a sequence dataframe (see details) and a phylogenetic tree}

\item{aa_ref}{amino acid reference sequence for alignment}

\item{nuc_ref}{nucleotide reference sequence for alignment}
}
\description{
Add ancestral state reconstruction to \code{tree_and_sequences}
}
\details{
The sequence dataframe in \code{tree_and_sequences} must contain columns \code{Isolate_unique_identifier} and \code{dna_sequence}. Other columns containing inforamtion about the sequences can also be present, and wil be preserved. Tip names in the tree in \code{tree_and_sequences} should correspond to names in the \code{Isolate_unique_identifier} column.
}
