% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence-package.R
\docType{package}
\name{convergence-package}
\alias{convergence}
\alias{convergence-package}
\title{convergence: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SamT123/convergence}
  \item Report bugs at \url{https://github.com/SamT123/convergence/issues}
}

}
\keyword{internal}
