% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{findBranch}
\alias{findBranch}
\title{Find branch with mutations}
\usage{
findBranch(
  tree_tibble,
  target_aa_substitutions,
  target_nt_mutations,
  allow_ambiguous = F
)
}
\arguments{
\item{tree_tibble}{tree_tibble}

\item{target_aa_substitutions}{aa substitutions which the branch should have}

\item{target_nt_mutations}{nt mutations which the branch should have}

\item{allow_ambiguous}{in cases where the aa and nt specification matches multiple branches, should an error be thrown (default), or should the branch with the largest number of descendant tips be used?}
}
\description{
Finds the branch where all of the target_aa_substitutions and target_nt_mutations are present, and at leaast occurs on the branch. Errors if the number of such branches is not exactly one.
}
