% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{getTimeIntervalSubstitutionRatios}
\alias{getTimeIntervalSubstitutionRatios}
\title{Title}
\usage{
getTimeIntervalSubstitutionRatios(
  tree_and_sequences,
  tree_info,
  time_boundaries,
  positions = 1:550,
  date_column = "estimated_date_nearest"
)
}
\arguments{
\item{tree_and_sequences}{tree_and_sequences list}

\item{tree_info}{tree_info from \code{getTreeSizeAndNucRates}}

\item{time_boundaries}{Boundaries between time intervals. The final interval will have boundaries (last_time_interval, Inf)}

\item{positions}{amino acid positions to analyse}

\item{date_column}{column to use for date filtering}
}
\description{
Title
}
