% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_size.R
\name{getTreeSizeAndNucRates}
\alias{getTreeSizeAndNucRates}
\title{Get information about a tree required for calculating convergence scores}
\usage{
getTreeSizeAndNucRates(tree_and_sequences)
}
\arguments{
\item{tree_and_sequences}{a list containing `tree` and `tree_tibble`

tree_size_ratios are the
overall_tree_size = overall_tree_size,
overall_nuc_rates = overall_nuc_rates,
fast_tree_size_fn = tree_size_fn}
}
\description{
Get information about a tree required for calculating convergence scores
}
