% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_size.R
\name{getTreeSizeAndNucRates}
\alias{getTreeSizeAndNucRates}
\title{Get information about a tree required for calculating convergence scores}
\usage{
getTreeSizeAndNucRates(
  tree_and_sequences,
  noise_model,
  excluded_positions = c(),
  use_fitted_rates = T,
  seed = NULL
)
}
\arguments{
\item{tree_and_sequences}{a list containing \code{tree} and \code{tree_tibble}}

\item{noise_model}{one of the 'noise_models'}

\item{excluded_positions}{positions to exclude for synonymous rate variation calculation}

\item{use_fitted_rates}{should the fitted mutation rates be used, rather than simply taking the mean of the observed values?}

\item{seed}{seed for rstan}

\item{n_iter}{iterations for rstan}

\item{n_chains}{chains for rstan}
}
\description{
Get information about a tree required for calculating convergence scores
}
