% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tree_functions.R
\name{make_cmaple_tree}
\alias{make_cmaple_tree}
\title{Wrapper for CMAPLE}
\usage{
make_cmaple_tree(
  sequences,
  tree_path,
  starting_tree_path = NULL,
  freeze_starting_tree = F
)
}
\arguments{
\item{sequences}{named character vector of DNA/RNA sequences}

\item{tree_path}{path to save tree}

\item{starting_tree_path}{path to starting tree containing a subset of the sequences}

\item{freeze_starting_tree}{should the relationships in the starting tree be preserved?}
}
\description{
Wrapper for CMAPLE
}
