% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{make_iq_tree}
\alias{make_iq_tree}
\title{Run IQ-TREE}
\usage{
make_iq_tree(
  alignment,
  wd,
  bb = NULL,
  nt = NULL,
  alrt = NULL,
  m = NULL,
  redo = FALSE,
  echo = FALSE,
  ...
)
}
\arguments{
\item{alignment}{DNA alignment to use for phylogenetic analysis. Must be matrix
(i.e., aligned sequences) of class DNAbin}

\item{wd}{Path to working directory. The alignment and IQ-TREE intermediate files
and results will be written here.}

\item{bb}{Optional; number of ultrafast bootstrap replicates to run.}

\item{nt}{Optional; number of cores to use. Set to "AUTO" to determine automatically.}

\item{alrt}{Optional; number of SH-aLRT tests to run.}

\item{m}{Optional; specify model. If no model is given, ModelTest will be run
to identify the best model for the data.}

\item{redo}{Logical; should the analysis be redone from scratch if output from
previous runs is present?}

\item{echo}{Logical; should STDERR be written to the screen?}

\item{...}{Other arguments not used by this function but used by
drake for tracking.}
}
\value{
Phylogenetic tree (list of class "phylo")
}
\description{
For details, see http://www.iqtree.org/doc/
}
\examples{
data(woodmouse)
# Rapid boot-strap tree with 1000 replicates on best-fitting model
tree <- iq_tree(woodmouse, tempdir(), bb = 1000, echo = TRUE)
plot(tree)
# Check the optimum number of cores to use for GTR+I+G model
iq_tree(tempdir(), woodmouse, m = "GTR+I+G", nt = "AUTO", echo = TRUE, redo = TRUE)
}
