% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimated_node_dates.R
\name{toChronumentalTree}
\alias{toChronumentalTree}
\title{Convert tree_and_sequences to a timetree using Chronumental}
\usage{
toChronumentalTree(
  tree_and_sequences,
  reference_strain,
  chronumental_path = NULL,
  n_steps = 10000,
  genome_size = NA
)
}
\arguments{
\item{tree_and_sequences}{a list containing \code{tree} and \code{tree_tibble}}

\item{reference_strain}{name of the strain to use as date reference (t=0)}

\item{chronumental_path}{path to chronumental binary}

\item{n_steps}{number of chronumental iterations}

\item{genome_size}{nt sequence length when branch lengths are in units of substitutions/site

Two types of node date are added: the collection date of the nearest tip, and the collection date of the nearest \emph{descendant} tip.}
}
\description{
Convert tree_and_sequences to a timetree using Chronumental
}
